v0.2.4 checks checksums for each P
Checks a K computed on the CPU against the final K on the GPU.  The
OpenCL code returns the last 24-bits of K along with the factor results.
If they do not match, or if any other computation error occurs, the
faulty test is re-run on the CPU.  CPU computations are also
double-checked in this way.

Most computation errors are thus non-fatal now.  But if at least 1/8 of
GPU computations fail in one run, or if the CPU fails 3 times in a row,
the test aborts.  Either of the above is more likely a bug than a
hardware problem.